refactor: consolidate AttributeDefinitionCreationException into Common\Exception and update related usages
