--- src/Domain/Identity/Exception/AttributeDefinitionCreationException.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PhpList\Core\Domain\Identity\Exception;
-
-use RuntimeException;
-
-class AttributeDefinitionCreationException extends RuntimeException
-{
-    private int $statusCode;
-
-    public function __construct(string $message, int $statusCode)
-    {
-        parent::__construct($message);
-        $this->statusCode = $statusCode;
-    }
-
-    public function getStatusCode(): int
-    {
-        return $this->statusCode;
-    }
-}
