+++ src/Domain/Identity/Service/Manager/AdminAttributeDefinitionManager.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace PhpList\Core\Domain\Identity\Service\Manager;
 
 use PhpList\Core\Domain\Common\Model\PaginatedResult;
-use PhpList\Core\Domain\Identity\Exception\AttributeDefinitionCreationException;
+use PhpList\Core\Domain\Common\Exception\AttributeDefinitionCreationException;
 use PhpList\Core\Domain\Identity\Model\AdminAttributeDefinition;
 use PhpList\Core\Domain\Identity\Model\Dto\AdminAttributeDefinitionDto;
 use PhpList\Core\Domain\Identity\Repository\AdminAttributeDefinitionRepository;
