+++ tests/Unit/Domain/Subscription/Service/Manager/AttributeDefinitionManagerTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace PhpList\Core\Tests\Unit\Domain\Subscription\Service\Manager;
 
 use PhpList\Core\Domain\Common\Model\AttributeTypeEnum;
-use PhpList\Core\Domain\Subscription\Exception\AttributeDefinitionCreationException;
+use PhpList\Core\Domain\Common\Exception\AttributeDefinitionCreationException;
 use PhpList\Core\Domain\Subscription\Model\Dto\AttributeDefinitionDto;
 use PhpList\Core\Domain\Subscription\Model\SubscriberAttributeDefinition;
 use PhpList\Core\Domain\Subscription\Repository\SubscriberAttributeDefinitionRepository;
